% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banmat.R
\name{banmat}
\alias{banmat}
\title{A Binary Ambiguous Nucleotide scoring Matrix (BAN Mat)}
\usage{
banmat()
}
\description{
Constructed based on NUC4.4 and designed for comparing ambiguous
sequences against "A", "T", "G", "C", and "N" containing sequences. 
Currently matches between ambiuous nucleotides are considered mismatch.
}
\details{
\code{banmat} returns a binary ambiguous nucleodide matrix.
}
\author{
Christopher Nobles, Ph.D.
}
