% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condense_intsites.R
\name{condense_intsites}
\alias{condense_intsites}
\title{Condense integration site GRanges to distinct genomic positions.}
\usage{
condense_intsites(sites.to.condense)

condense_intsites(sites.to.condense, grouping = NULL,
return.abundance = FALSE, method = "fragLen", replicates = "replicates")
}
\arguments{
\item{sites.to.condense}{a GRanges object where each row represents a genomic
range of reference genome DNA.}

\item{grouping}{metadata column name (input as character) which designates
which genomic ranges belong to which sample.}

\item{return.abundance}{logical that will subject the data to abundance
calculations, default is FALSE.}

\item{method}{either "fragLen" (default) or "estAbund" to be passed to
determine_abundance function. method = "fragLen" determines the abundance by
the number of unique genomic ranges widths while the method = "estAbund"
utilizes the sonicLength package.}

\item{replicates}{character name of the column carrying the replicate
information, relevant for abundance calculations and irrelent after using the
this function.}

\item{quiet}{logical indicating if output message should be displayed.}
}
\description{
Given a GRanges object representing the genomic ranges of
sequenced DNA flanking the viral/vector integrating site, this function
returns only the distinct integration site positions. Further, using the
options for return.abundance and method, the function can return abundance
or estimated abundance information based on fragment length (see sonicLength
package). Initial GRanges object can carry multiple samples which can be
separated and analyzed independently by using the grouping option.
}
\details{
\code{condense_intsites} condenses or collapses GRanges data to distinct
integration positions in the reference genome.
}
\examples{
gr <- gintools:::generate_test_granges()
std.gr <- standardize_sites(gr)
condense_intsites(std.gr)

condense_intsites(std.gr, return.abundance = TRUE)

}
\author{
Christopher Nobles, Ph.D.
}
