% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_standardization_methods.R
\name{.standardizeSites}
\alias{.standardizeSites}
\title{Wrapper around .clusterSites() for standardizing integration site positions
given integration site distributions.}
\usage{
.standardizeSites(unstandardizedSites)
}
\description{
This function coerces GRanges objects of integration site data
into a format to work with the .clusterSites() function for standardizing
integration sites to a single nucleotide position in the reference geneome
given a set or distribution of integration sites.
}
\author{
Bushman Lab
}
