% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_posid.R
\name{generate_posid}
\alias{generate_posid}
\title{Generate position IDs.}
\usage{
generate_posid(sites = NULL)

generate_posid(seqnames = NULL, strand = NULL, start = NULL, end = NULL, delim = "")
}
\arguments{
\item{sites}{a GRanges object where each row represents one integrated
element.}

\item{seqnames}{a character vector of chromosomes for integrated elements,
i.e. chr1, chr3, chrX.}

\item{strand}{orientation or strand on which the integrated element
is on (+, -, *)}

\item{start}{the lower numerical position of the integrated element,
considered the start for "+" integrated elements.}

\item{end}{the greater numerical position of the integrated element,
considered the end for "+" integrated elements.}

\item{delim}{a single character to delimit the data within the string.}
}
\description{
Given a set of genomic integration positions, the
function generates a character vector of integrated positions in the
following format: chr(+/-)pos, where chr is the chromosome, (+/-) designates
the orientation of the integration, and pos is the position of the integrated
element.
}
\details{
\code{generate_posid} generates a character vector of position IDs
}
\examples{
chr <- c("chr1", "chr3", "chrX")
strands <- c("+", "-", "+")
starts <- c(900231, 13254892, 603292)
ends <- c(900431, 13255292, 603592)
ranges <- IRanges(start = starts, end = ends)
gr <- GRanges(seqnames = chr, ranges = ranges, strand = strands)

generate_posid(sites = gr)

generate_posid(seqnames = chr, strand = strands, start = starts, end = ends)

}
\author{
Christopher Nobles, Ph.D.
}
