% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_granges.R
\name{generate_test_granges}
\alias{generate_test_granges}
\title{Generate a test GRanges object}
\usage{
gintools:::generate_test_granges(n_sites = 5, n_reads_p_site = 20, site_range = 1:1000, read_width_range = 30:100)
}
\arguments{
\item{n_sites}{number of sites to generate}

\item{n_reads_p_site}{number of reads per site}

\item{site_range}{numeric vector to sample from for postion of the integration
sites}

\item{read_width_range}{range of widths to sample from for each read}

\item{stdev}{numeric, standard deviation of the integration site distribution
around the selected positions}

\item{positions}{a numeric vector of preselected positions.}
}
\description{
Generate a test GRanges object
}
\details{
Varies the integration start position in with a normal distribution
around the random integration site and then assignes a random uniform
distribution of widths for the sites.
}
\examples{
gintools:::generate_test_granges()

}
\author{
Christopher Nobles, Ph.D.
}
