% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_clusters.R
\name{graph_clusters}
\alias{graph_clusters}
\title{Generate a graph from integrations sites within a specified genomic gap
distance}
\usage{
graph_clusters(sites, gap)
}
\arguments{
\item{sites}{a GRanges object where each row is a single integration site or
range.}

\item{gap}{an integer specifying the distance to consider between sites to
call an edge.}
}
\description{
Given a set of integration sites, this function generates a
partial undirected graph of integration sites where nodes are the specific
sites (one row from the GRanges object input) and edges mark which sites are
within the gap distance from eachother. To generate the partial graph, nodes
are broken into two groups, axil nodes and nodes. Axil nodes are the first
occurances (or left most occurance) of the unique integration sites
(chromsome + strand + position) which cluster together. Nodes are all other
occurances, which are always connected to one axil node by definition.
}
\details{
\code{graph_clusters} generates a partial undirected graph connecting each
integration sites within a specified genomic gap distance.
}
\examples{
sites <- GRanges(
  seqnames = rep("chr1", 7),
  ranges = IRanges(
    start = c(50, 49, 44, 50, 50, 50, 60),
    width = seq(20, 26, 1)),
  strand = rep("+", 7))

graph <- graph_clusters(sites, gap = 5L)

plot.igraph(graph)

}
\author{
Christopher Nobles, Ph.D.
}
