% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_intsite_positions.R
\name{normalize_intsite_positions}
\alias{normalize_intsite_positions}
\title{Normalize integration site positions between samples}
\usage{
normalize_intsite_positions(sites)

normalize_intsite_positions(sites, gap = 0L)
}
\arguments{
\item{sites}{GRanges object where each row represents one integration range
or site.}

\item{gap}{integer for the distance between sites to consider.}
}
\description{
Integration sites across multiple samples may be identified at
slightly different nucleotide positions. This function identifies sites
within the gap distance from eachother and normalized their position to the
most frequent position. Similar to standardizing the integration sites, but
not as logic heavy.
}
\details{
\code{normalize_intsite_positions} identifies sites within a gap distance and
assigns each of the sites in the group the most frequent position as well as
a positionID (pos.clus).
}
\examples{

gr <- GRanges(
  seqnames = rep("chr1", 10),
  ranges = IRanges(
    start = rnorm(10, 20, 1),
    width = rep(15, 10)),
  strand = rep("+", 10))

normalize_intsite_positions(gr)

normalize_intsite_positions(gr, gap = 1L)

}
\author{
Christopher Nobles, Ph.D.
}

