% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_calcs.R
\name{pop_calcs}
\alias{calc_clonality}
\alias{calc_entropy}
\alias{calc_gini}
\alias{calc_shannon}
\alias{pop_calcs}
\title{Calculation for descriptors of populations}
\usage{
pop_calcs(x, calc)

calc_shannon(x, base = exp(1))

calc_gini(x)

calc_entropy(x)

calc_clonality(x)
}
\arguments{
\item{x}{numeric a vector of abundances or proportions / frequencies.}

\item{calc}{character one of the various population calcuations. Choices are:
"shannon", "gini", "entropy", or "clonality".}
}
\description{
Given a population of integration sites, TCR/IGH sequences,
etc., a numerical value can be used to describe the diveristy (shannon index
or entropy), or the clonality (gini or clonality). These numerical values
allow for a comparison to be made between populations. Entropy and Clonality
are factors used by Adaptive Biotechnologies to describe their TCR / IGH
sequence data.
}
\details{
\code{pop_calcs} Calculations for describing features of populations.
}
\section{Functions}{
\itemize{
\item \code{calc_shannon}: Calculate Shannon Diversity.

\item \code{calc_gini}: Calculate gini index.

\item \code{calc_entropy}: Calculate Entropy as defined by Adaptive Biotech.

\item \code{calc_clonality}: Calcuate Clonality as defined by Adaptive Biotech.
}}
\examples{
x <- sample(1:100, 50, replace = TRUE)

calc_shannon(x)
calc_gini(x)
calc_entropy(x)
calc_clonality(x)

}
\author{
Christopher Nobles, Ph.D.
}

