% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine_breakpoints.R
\name{refine_breakpoints}
\alias{refine_breakpoints}
\title{Refine or resolve sonic break points within a dataset}
\usage{
refine_breakpoints(sites)

refine_breakpoints(sites, count = "counts", min.gap = 1L, sata.gap = 3L)
}
\arguments{
\item{sites}{GRanges object. Integration site ranges to adjust.}

\item{counts}{character string. Name of column holding range count
information (ie. read counts). If not supplied, assume count of 1 for each
row in sites.}

\item{min.gap}{integer minimum gap to consider combining break points.}

\item{sata.gap}{integer maximum distance to consider combining break points.}
}
\description{
Given a GRanges object ....
}
\details{
\code{refine_breakpoints} returns a GRanges object where the sonic break
point positions have been adjusted based on the dataset and counts.
}
\examples{
gr <- gintools:::generate_test_granges()

refine_breakpoints(gr)

}
\author{
Christopher L. Nobles, Ph.D.
}
