% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_format.R
\name{scan_format}
\alias{scan_format}
\title{Format a set of GRanges objects for geneRxCluster}
\usage{
scan_format(query, subject, grouping = NULL)
}
\arguments{
\item{query}{GRanges object consistent with integration sites, will be
associated with grp = 0.}

\item{subject}{GRanges object consistent with integration sites, will be
associated with grp = 1.}

\item{grouping}{character name of metadata column in GRanges object to be
used for defining groups. Groups will be made unique independently, meaning
if the same event is observed in two different groups, they are both kept,
but if they occur in the same group, they will be made one.}
}
\description{
Format a set of GRanges objects for geneRxCluster
}
\details{
Formats a GRanges object for use with the geneRxCluster
functions, specifically the scan statistics of gRxCluster.
}
\examples{
gr1 <- gintools:::generate_test_granges()
gr2 <- gintools:::generate_test_granges()
scan_format(gr1, gr2)

}
\author{
Christopher Nobles, Ph.D.
}
