% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serial_cluster.R
\name{serial_cluster}
\alias{serial_cluster}
\title{Generate clusterIDs for integration sites given multiple gap distances.}
\usage{
serial_cluster(sites, gaps = c(0L, 1L, 2L))
}
\arguments{
\item{sites}{a GRanges object where each row is a different integration site
or range.}

\item{gaps}{an integer vector of any length specifying the nucleotide
distance allowable between sites to call in the same group or cluster.}
}
\description{
Given a GRanges object, this function identifies groups or
clusters of integration sites within specified distances. The output is of
the same length and order of the input, and clusterID information is appended
to the metadata columns.
}
\details{
\code{serial_cluster} returns a GRanges object of the same length and order
as input with additional metadata columns specifying the group or clusterID
of given nt windows.
}
\examples{
sites <- GRanges(
  seqnames = rep("chr1", 7),
  ranges = IRanges(
    start = c(50, 49, 44, 50, 50, 50, 60),
    width = seq(20, 26, 1)),
  strand = rep("+", 7))

serial_cluster(sites, gaps = c(1L, 5L, 10L))

}
\author{
Christopher Nobles, Ph.D.
}
