% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinks.R
\name{sinks}
\alias{sinks}
\title{Identify sinks in a directed graph}
\usage{
sinks(graph)
}
\arguments{
\item{graph}{a directed graph (igraph).}
}
\description{
From a directed graph, this function returns all sink nodes, or
nodes which only act as tail nodes and not as head nodes.
}
\details{
\code{sinks} returns a numerical vector of sink nodes.
}
\examples{
g <- make_graph(edges = c(1,2, 2,3, 1,3), directed = TRUE)
sinks(g)

}
\author{
Christopher Nobles, Ph.D.
}
