% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{sources}
\alias{sources}
\title{Identify sources in a directed graph}
\usage{
sources(graph)
}
\arguments{
\item{graph}{a directed graph (igraph).}
}
\description{
From a directed graph, this function returns all source nodes, or
nodes which only act as head nodes and not as tail nodes.
}
\details{
\code{sources} returns a numerical vector of source nodes.
}
\examples{
g <- make_graph(edges = c(1,2, 2,3, 1,3), directed = TRUE)
sources(g)

}
\author{
Christopher Nobles, Ph.D.
}
