% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_sites.R
\name{standardize_sites}
\alias{standardize_sites}
\title{Standardize genomic site positions within a dataset}
\usage{
standardize_sites(input.sites)

standardize_sites(input.sites, counts.col = "counts", min.gap = 1L, sata.gap = 5L)
}
\arguments{
\item{input.sites}{GRanges object of sites to standardize.}

\item{counts.col}{character string. Name of column holding range count
information (ie. read counts). If not supplied, assume count of 1 for each
row in sites.}

\item{min.gap}{integer minimum gap to consider combine sites.}

\item{sata.gap}{integer maximum distance to consider combining sites.}

\item{details}{logical If TRUE, data columns will be appended to the metadata
of the GRanges object noting the original and adjusted position for the site.
FALSE by default.}

\item{quiet}{logical during processing, should messages be suppressed which
report findings? TRUE by default.}
}
\description{
Given a GRanges object ...
}
\details{
\code{standardize_sites} returns a GRanges object where the site positions
have been standardized with sites within the gap distance.
}
\examples{
gr <- gintools:::generate_test_granges()

standardize_sites(gr)

}
\author{
Christopher Nobles, Ph.D.
}
