% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_clones.R
\name{track_clones}
\alias{track_clones}
\title{Identifies or tracks integration sites through multiple GRanges objects.}
\usage{
track_clones(sites.list)

track_clones(sites.list, gap = 5L, track.origin = TRUE)
}
\arguments{
\item{sites.list}{a GRangesList object containing sets of integration sites
or ranges (one per row).}

\item{gap}{an integer designating the nucleotide distance or window for which
to group integration sites.}

\item{track.origin}{logical for whether to append information regarding the
origin of the integration site. Use if there is no sample information in the
metadata columns of the input GRanges objects.}

\item{quiet}{logical for whether or not to message to the terminal processing
findings. True by default.}
}
\description{
Given a GRangesList of integration site sets, this function will
return a GRangesList (separated by position ID) of all sites shared within
the gap distance between the sets. Using the track.origin option
(default = TRUE) will append a column to the metadata of the GRanges objects
designating which set they originated in. If the GRangesList has names, these
designations will be the names of the list items. Note: as the output is a
GRangesList separated by position ID, some list objects may appear to come
only from one origin set, but they will overlap within the gap distance to
another list object from a different origin.
}
\details{
\code{track_clones} returns a GRangesList of integration sites shared between
multiple GRanges objects.
}
\examples{
gr <- GRanges(
  seqnames = rep("chr1", 100),
  ranges = IRanges(
    start = 1:100,
    width = sample(30:100, 100, replace = TRUE)),
  strand = rep(c("+", "-"), 50))

gr1 <- sample(gr, 50)
gr2 <- sample(gr, 50)

grl <- GRangesList(gr1, gr2)

track_clones(grl, gap = 0L, track.origin = TRUE)

}
\author{
Christopher Nobles, Ph.D.
}
