% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_demultiplexed_sequences.R
\name{write_demultiplexed_sequences}
\alias{write_demultiplexed_sequences}
\title{Write demultiplexed sequences to fasta/q files in an output director}
\usage{
write_demultiplexed_sequences(read_file_path, multiplexed_data, out_folder,
  type, read_name_pattern = "[\\\\w\\\\:\\\\-\\\\+]+", compress = TRUE)
}
\arguments{
\item{read_file_path}{path to sequence file (fasta/q) to be demultiplexed.}

\item{out_folder}{path to output directory.}

\item{type}{character identifying the read class (R1, I1, I2, R2).}

\item{read_name_pattern}{regular expression to be applied to the sequence 
names for identification. Default: "[\\w\\:\\-\\+]+".}

\item{compress}{logical indicating if the output files should be gzip 
compressed (.gz, default) or uncompressed.}
}
\description{
This function will read in a sequence file, parse it given 
`multiplexed_data` information and write the parsed files to an output 
directory. Specify the `type` as "fasta" or "fastq" and give a sequence file
path (`read_file_path`) and the output folder path (`out_folder`). Data can 
either be comressed (into .gz format) or uncompressed.
}
\details{
\code{write_demultiplexed_sequences} will write fasta/q files to an output 
directory when supplied with grouping information (`multiplexed_data`).
}
\author{
Christopher Nobles, Ph.D.
}
