% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_intsites.R
\name{standardize_intsites}
\alias{standardize_intsites}
\title{Standardize genomic integration site positions within a dataset}
\usage{
standardize_intsites(sites)

standardize_intsites(sites, min.gap = 1L, sata.gap = 5L)
}
\arguments{
\item{sites}{GRanges Object integration sites to standardize.}

\item{min.gap}{integer minimum gap to consider combine integration sites.}

\item{sata.gap}{integer maximum distance to consider combining integration
sites.}
}
\description{
Given a GRanges object ...
}
\details{
\code{standardize_intsites} returns a GRanges object where the integration
site positions have been standardized with sites within the gap distance.
}
\examples{
gr <- gintools:::.generate_test_granges()

standardize_intsites(gr)

}
\author{
Christopher Nobles, Ph.D.
}

