% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_intsites.R
\name{standardize_intsites}
\alias{standardize_intsites}
\title{gr <- GRanges(
  seqnames = rep("chr1", 100),
  ranges = IRanges(
    start = sample(1:200, 100, replace = TRUE),
    width = sample(30:200, 100, replace = TRUE)),
  strand = rep("+", 100))}
\usage{
standardize_intsites(unstandardized.sites, std.gap = 1L,
  standardize_breakpoints = FALSE, get.analysis.data = FALSE,
  calc.cluster.stats = FALSE)
}
\description{
gr <- GRanges(
  seqnames = rep("chr1", 100),
  ranges = IRanges(
    start = sample(1:200, 100, replace = TRUE),
    width = sample(30:200, 100, replace = TRUE)),
  strand = rep("+", 100))
}

