% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_intsites.R
\name{standardize_intsites}
\alias{standardize_intsites}
\title{Standardize genomic integration site positions within a dataset}
\usage{
standardize_intsites(sites)

standardize_intsites(sites, std.gap = 1L, standardize_breakpoints = FALSE,
get.analysis.data = FALSE, calc.cluster.stats = FALSE)
}
\arguments{
\item{sites}{}
}
\description{
Given a GRanges object ...
}
\details{
\code{standardize_intsites} returns a GRanges object where the integration
site positions have been standardized with sites within the gap distance.
}
\examples{
gr <- gintools:::.generate_test_granges()

standardize_intsites(gr)

}
\author{
Christopher Nobles, Ph.D.
}

